Load WpcomChecklist asynchronously in stats page.
WpcomChecklist loads a sizeable amount of code, but it's only used
in sites that qualify and haven't followed the checklist to the end.

Instead of statically including and instancing the component every
time, this PR instead extracts some minimal exports to determine
whether it makes sense to display the WpcomChecklist banner. Only
if these checks are passed does the WpcomChecklist component then
get asynchronously loaded and instanced.